Fix circular dependency between adapter and types, add default behavior to the TeradataExpressionAdapter for when a visit method is not defined on the left operand, and fix for the binary types test case relying on get_columns to be ordered a certain way.
